Moved mesos-java code into the mesos-hadoop tree
