--- 1.1/README.md
@@ -1,1 +1,0 @@
-# mesos-java
